Add support for the process per thread run strategy in libtest.
This essentially just adds a (hidden) option to continue running the child process past the provided function.
The process per thread run strategy is used by libtest when `-Zpanic_abort_tests` is given to Cargo (this enables running tests with `-Cpanic=abort`) and `force_run_in_process` is not passed to the libtest runner.